refactor(theme): move useResolvedTheme to src/hooks beside the app theme hooks
Review: the hook is not analytics-specific — it resolves the ThemeProvider's
.dark class for any JS code that must branch on effective theme. Chart-token
lookup (getChartColors) stays feature-local in analytics/lib/theme.ts.
